requirements.txt
openai
googlesearch-python
requests
beautifulsoup4
